hearch-v1: fixed issues and functionality added

+++ Hearch/app/src/main/java/com/example/myapplication/model/RETIREMENT.kt
@@ -4,5 +4,5 @@ import androidx.databinding.ObservableDouble
 import androidx.databinding.ObservableInt
 import java.time.Month
 
-data class RETIREMENT(var ageNow:Int,var retirementAge: Int,var lifeExpectancy:Int,
-var currentCorpus:Double,var CurrentExpense:Double,var expectedReturn:Double,var sipMonth: ObservableDouble)
+data class RETIREMENT(var ageNow:Long,var retirementAge: Long,var lifeExpectancy:Long,
+var currentCorpus:Long,var CurrentExpense:Long,var expectedReturn:Double,var sipMonth: ObservableDouble)
